refactor IR packages
