+++ src/main/kotlin/edu/itmo/ilang/ir/model/Body.kt
@@ -1,4 +1,4 @@
-package edu.itmo.ilang.ir
+package edu.itmo.ilang.ir.model
 
 sealed interface BodyEntry : IrEntry
 data class Body(
