+++ src/main/kotlin/edu/itmo/ilang/ir/model/Program.kt
@@ -1,4 +1,4 @@
-package edu.itmo.ilang.ir
+package edu.itmo.ilang.ir.model
 
 sealed interface IrEntry
