+++ src/main/kotlin/edu/itmo/ilang/semantic/analysis/Analyser.kt
@@ -1,6 +1,6 @@
 package edu.itmo.ilang.semantic.analysis
 
-import edu.itmo.ilang.ir.Program
+import edu.itmo.ilang.ir.model.Program
 
 interface Analyser {
     fun analyse(program: Program)
