+++ src/main/kotlin/edu/itmo/ilang/semantic/checkers/BodyEntriesChecker.kt
@@ -1,6 +1,7 @@
 package edu.itmo.ilang.semantic.checkers
 
 import edu.itmo.ilang.ir.*
+import edu.itmo.ilang.ir.model.Program
 
 interface BodyEntriesChecker : Checker, BodyEntriesProcessor {
     override fun check(program: Program) {
