+++ src/main/kotlin/edu/itmo/ilang/semantic/checkers/Checker.kt
@@ -1,6 +1,6 @@
 package edu.itmo.ilang.semantic.checkers
 
-import edu.itmo.ilang.ir.Program
+import edu.itmo.ilang.ir.model.Program
 
 interface Checker {
     fun check(program: Program)
