+++ src/main/kotlin/edu/itmo/ilang/semantic/checkers/TypeChecker.kt
@@ -1,6 +1,6 @@
 package edu.itmo.ilang.semantic.checkers
 
-import edu.itmo.ilang.ir.Program
+import edu.itmo.ilang.ir.model.Program
 
 class TypeChecker : Checker {
     override fun check(program: Program) {
