+++ src/main/kotlin/edu/itmo/ilang/semantic/transformations/Transformer.kt
@@ -1,6 +1,6 @@
 package edu.itmo.ilang.semantic.transformations
 
-import edu.itmo.ilang.ir.Program
+import edu.itmo.ilang.ir.model.Program
 
 interface Transformer {
     fun transform(program: Program)
