+++ src/test/kotlin/tests/analysis/TestDeadCodeEliminator.kt
@@ -1,6 +1,6 @@
 package tests.analysis
 
-import edu.itmo.ilang.ir.*
+import edu.itmo.ilang.ir.model.*
 import edu.itmo.ilang.semantic.transformations.DeadCodeEliminator
 import generateIr
 import org.junit.jupiter.api.Test
